new command .auto-run, ported scripts, updated documentation
