--- sample_scripts/lua_lib_sample.lua
@@ -1,3 +1,0 @@
-zelph = require("zelph")
-
-local R = zelph.fact("A", "is a", "B")
