--- sample_scripts/notes.md
@@ -1,47 +1,0 @@
-Q23444  is a  Q1075 <= (Q23445  is opposite of  Q23444), (Q23445  is a  Q1075)
-Q23444  is subclass of  Q40890790 <= (Q23445  is opposite of  Q23444), (Q23445  is subclass of  Q40890790)
-Q23444  is opposite of  Q23445 <= Q23445  is opposite of  Q23444
-! <= (Q23444  is a  Q1075), (Q23444  is subclass of  Q40890790)
-
-
-white  is a  color <= (black  is opposite of  white), (black  is a  color)
-white  is subclass of  absorbed or reflected object light <= (black  is opposite of  white), (black  is subclass of  absorbed or reflected object light)
-white  is opposite of  black <= black  is opposite of  white
-! <= (white  is a  color), (white  is subclass of  absorbed or reflected object light)
-
-
-white  is a  color <= (black  is opposite of  white), (black  is a  color)
-white  is subclass of  absorbed or reflected object light <= (black  is opposite of  white), (black  is subclass of  absorbed or reflected object light)
-white  is opposite of  black <= black  is opposite of  white
-! <= (white  is a  color), (white  is subclass of  absorbed or reflected object light)
-
-
-Q458 P1151 Q3249016
-European Union  topic's main Wikimedia portal  Portal:European Union
-
-469 4 948 - 4806982390444457984
-981 4 950 - 4806982390444457984  (4611686018428365476 ??)
-
-4: instance of
-469: Flag of Italy
-981: flag of the United Nations
-948: national flag
-950: flag
-
-
-    vec: {948}
-		c1	4806466702311161856	unsigned __int64 0x42B4000000000000
-		c3	515670953426944	unsigned __int64        0x1D50000000000
-
-		c1	4807029652264583168	unsigned __int64 0x42B6000000000000
-		c3	1078620906848256	unsigned __int64      0x3D50000000000
-
-		c1	0x458d000000000000	unsigned __int64
-		c2	0x000000004d000000	unsigned __int64
-		c3	0x00000fd600000000	unsigned __int64
-		result	0x458d0fd64d000000	unsigned __int64
-
-		c1	0x458d000000000000	unsigned __int64
-		c2	0x000000014d000000	unsigned __int64
-		c3	0x00000fd700000000	unsigned __int64
-		result	0x458d0fd64d000000	unsigned __int64
