--- sample_scripts/simple.zph
@@ -1,4 +1,0 @@
-
-X is opposite of Y => Y is opposite of X
-white is opposite of black
-.dot white 10 test.dot
