--- sample_scripts/test.zph
@@ -1,3 +1,0 @@
-X is a Y  => X ~ Y
-is is a ->
-has part is a ->
