--- sample_scripts/test2.zph
@@ -1,38 +1,0 @@
-uint p,q,n
-
-if there is no q in [2..x-1] with p mod q=0 then p is prime
-if p mod (q*n)=0 then p mod q=0
-
-if n<=2 then fib(n)=1 else fib(n)=fib(n-1)+fib(n-2)
-
-window1 = window("Test", size(500,500))
-
-frame1 = frame(500,500)
-
-plot color(50,50,50) at 100,150 on frame1
-
-show frame1 at 0,0 on window1
-
-r,g,b is red if r>>g and r>>b
-r,g,b is yellow if r~g and r>>b
-
-sprite1 = circle(50,50)
-fill sprite1 with red at 50,50
-
-show sprite1 at 250,250 on window1
-
-button1 is button
-
-show button1 at 300,300 on window1
-
-if button1 was clicked then move sprite1 by 3,3 on window1
-
-define plot color at coord on frame:
-  # use cv::Mat
-
-
-button is a filled rectangle
-
-thing is filled if color of != color of background of thing
-
-background of thing is average color of
