+++ sample_scripts/wikidata-disjointness-violations-var-length1_3.zph
@@ -22,24 +22,24 @@
 # =============================================================================
 
 # Depth 1+1
-K "P279" e1, K "P279" e2 => !
+(*{(K "P279" e1) (K "P279" e2)} ~ conjunction) => !
 
 # Depth 1+2 and 2+1
-K "P279" e1, K "P279" Y, Y "P279" e2 => !
-K "P279" X, X "P279" e1, K "P279" e2 => !
+(*{(K "P279" e1) (K "P279" Y) (Y "P279" e2)} ~ conjunction) => !
+(*{(K "P279" X) (X "P279" e1) (K "P279" e2)} ~ conjunction) => !
 
 # Depth 2+2
-K "P279" X, X "P279" e1, K "P279" Y, Y "P279" e2 => !
+(*{(K "P279" X) (X "P279" e1) (K "P279" Y) (Y "P279" e2)} ~ conjunction) => !
 
 # Depth 1+3 and 3+1
-K "P279" e1, K "P279" A, A "P279" B, B "P279" e2 => !
-K "P279" A, A "P279" B, B "P279" e1, K "P279" e2 => !
+(*{(K "P279" e1) (K "P279" A) (A "P279" B) (B "P279" e2)} ~ conjunction) => !
+(*{(K "P279" A) (A "P279" B) (B "P279" e1) (K "P279" e2)} ~ conjunction) => !
 
 # Depth 2+3 and 3+2
-K "P279" X, X "P279" e1, K "P279" A, A "P279" B, B "P279" e2 => !
-K "P279" A, A "P279" B, B "P279" e1, K "P279" Y, Y "P279" e2 => !
+(*{(K "P279" X) (X "P279" e1) (K "P279" A) (A "P279" B) (B "P279" e2)} ~ conjunction) => !
+(*{(K "P279" A) (A "P279" B) (B "P279" e1) (K "P279" Y) (Y "P279" e2)} ~ conjunction) => !
 
 # Depth 3+3
-K "P279" A, A "P279" B, B "P279" e1, K "P279" C, C "P279" D, D "P279" e2 => !
+(*{(K "P279" A) (A "P279" B) (B "P279" e1) (K "P279" C) (C "P279" D) (D "P279" e2)} ~ conjunction) => !
 
 .run-md disjointness-violations-var-length1-3
