+++ sample_scripts/wikidata-only-deductions.zph
@@ -24,30 +24,31 @@
 # The following facts are not part of wikidata:
 "has quality" ~ "transitive relation"
 
-X "is facet of" Y, Y "has part" P => X "has part" P
-X "is facet of" Y, Y "is part of" P => X "is part of" P
-X "is facet of" Y, Y "has quality" Q => X "has quality" Q
+(*{(X "is facet of" Y) (Y "has part" P)}    ~ conjunction) => (X "has part" P)
+(*{(X "is facet of" Y) (Y "is part of" P)}  ~ conjunction) => (X "is part of" P)
+(*{(X "is facet of" Y) (Y "has quality" Q)} ~ conjunction) => (X "has quality" Q)
 
 # The following fact is not part of wikidata. Wikidata only includes the fact "is subclass of" "subject item of this property" "is for example"
 "is for example"  "is inverse of" ~
 
-R ~ "transitive relation", X R Y, Y R Z => X R Z
-P ~ "transitive relation", P "is inverse of" Q => Q ~ "transitive relation"
-X ~ K, K "is subclass of" U => X ~ U
+(*{(R ~ "transitive relation") (X R Y) (Y R Z)} ~ conjunction) => (X R Z)
+(*{(P ~ "transitive relation") (P "is inverse of" Q)} ~ conjunction) => (Q ~ "transitive relation")
+(*{(X ~ K) (K "is subclass of" U)} ~ conjunction) => (X ~ U)
 
-X "has quality" E,   E ~ K                => X "has quality" K
-X "has quality" E,   E "is subclass of" K  => X "has quality" K
-K "has quality" E,   X ~ K                => X "has quality" E
-K "has quality" E,   X "is subclass of" K  => X "has quality" E
-X "has part" P,      P ~ K                => X "has part" K
-K "has part" P,      X "is subclass of" K  => X "has part" P
+(*{(X "has quality" E) (E ~ K)}                ~ conjunction) => (X "has quality" K)
+(*{(K "has quality" E) (X ~ K)}                ~ conjunction) => (X "has quality" E)
+(*{(X "has part" P)    (P ~ K)}                ~ conjunction) => (X "has part" K)
+(*{(X "has quality" E) (E "is subclass of" K)} ~ conjunction) => (X "has quality" K)
+(*{(K "has quality" E) (X "is subclass of" K)} ~ conjunction) => (X "has quality" E)
+(*{(K "has part" P)    (X "is subclass of" K)} ~ conjunction) => (X "has part" P)
 
-X "is opposite of" Y, X ~ K              => Y ~ K
-X "is opposite of" Y, X "is subclass of" K => Y "is subclass of" K
-X "is inverse of" Y,  X ~ K              => Y ~ K
-X "is inverse of" Y,  X "is subclass of" K => Y "is subclass of" K
+(*{(X "is opposite of" Y) (X ~ K)}                ~ conjunction) => (Y ~ K)
+(*{(X "is inverse of" Y)  (X ~ K)}                ~ conjunction) => (Y ~ K)
+(*{(X "is opposite of" Y) (X "is subclass of" K)} ~ conjunction) => (Y "is subclass of" K)
+(*{(X "is inverse of" Y)  (X "is subclass of" K)} ~ conjunction) => (Y "is subclass of" K)
 
-X "is opposite of" Y        => Y "is opposite of" X
-X "is inverse of" Y         => Y "is inverse of" X
-R "is opposite of" S, X R Y => Y S X
-R "is inverse of" S,  X R Y => Y S X
+(X "is opposite of" Y) => (Y "is opposite of" X)
+(X "is inverse of" Y)  => (Y "is inverse of" X)
+
+(*{(R "is opposite of" S) (X R Y)} ~ conjunction) => (Y S X)
+(*{(R "is inverse of" S)  (X R Y)} ~ conjunction) => (Y S X)
