--- sample_scripts/wikidata-var.zph
@@ -1,75 +1,0 @@
-.lang zelph
-
-.name !                wikidata Q363948
-.name ~                zelph "is a"
-.name ~                wikidata P31
-.name "is subclass of" wikidata P279
-.name =>               wikidata Q7881229
-.name ->               wikidata Q18616576
-.name "is part of"     wikidata P361
-.name "has part"       wikidata P527
-.name "is opposite of" wikidata P461
-.name "is inverse of"  wikidata P1696
-.name "has quality"    wikidata P1552
-.name "is for example" wikidata Q21514624
-.name "transitive relation" wikidata Q18647515
-
-# The following facts are part of wikidata:
-#"is subclass of" is a transitive relation
-#"has part"       is a transitive relation
-#"is part of"     is inverse of "has part"
-
-# The following facts are not part of wikidata:
-"has quality" is a transitive relation
-
-# The following fact is not part of wikidata. Wikidata only includes the fact "is subclass of" "subject item of this property" "is for example"
-"is for example"  is inverse of "is a"
-
-R is a transitive relation, X R Y, Y R Z => X R Z
-P is a transitive relation, P is inverse of Q => Q is a transitive relation
-X is a K, K is subclass of U => X is a U
-
-X has quality E,   E is a K            => X has quality K
-X has quality E,   E is subclass of K  => X has quality K
-K has quality E,   X is a K            => X has quality E
-K has quality E,   X is subclass of K  => X has quality E
-X has part P,      P is a K            => X has part K
-K has part P,      X is subclass of K  => X has part P
-
-X is opposite of Y, X is a K           => Y is a K
-X is opposite of Y, X is subclass of K => Y is subclass of K
-X is inverse of Y,  X is a K           => Y is a K
-X is inverse of Y,  X is subclass of K => Y is subclass of K
-
-X is opposite of Y        => Y is opposite of X
-X is inverse of Y         => Y is inverse of X
-R is opposite of S, X R Y => Y S X
-R is inverse of S,  X R Y => Y S X
-
-X is opposite of Y, A has quality X, A has quality Y => !
-X is inverse of Y,  A has quality X, A has quality Y => !
-X is opposite of Y, A has part X, A has part Y => !
-X is inverse of Y,  A has part X, A has part Y => !
-
-X is opposite of Y, A is a X,           A is a Y           => !
-X is opposite of Y, A is subclass of X, A is subclass of Y => !
-X is inverse of Y,  A is a X,           A is a Y           => !
-X is inverse of Y,  A is subclass of X, A is subclass of Y => !
-
-X has quality E, X is a E           => !
-X has quality E, X is subclass of E => !
-X has quality E, E is a X           => !
-X has quality E, E is subclass of X => !
-X has quality E, E has part X       => !
-
-X has part E, X is a E           => !
-X has part E, X is subclass of E => !
-X has part E, E is a X           => !
-X has part E, E is subclass of X => !
-
-# The following contradiction requires that X cannot be at the same time an instance and a subclass:
-X is a A, X is subclass of B => !
-
-A is a B, B is a A                     => !
-A is subclass of B, B is subclass of A => !
-A is a B, B is subclass of A           => !
